+++ endpoints/v2/test/test_v2auth.py
@@ -254,6 +254,17 @@ def get_robot_password(username):
             "private",
             False,
         ),
+        # Proxy organization repository, org creator, repository does not exist
+        (
+            "repository:proxyorg/idonotexist:pull",
+            "member",
+            "password",
+            200,
+            ["proxyorg/idonotexist:pull"],
+            True,
+            "private",
+            False,
+        ),
         # Organization repository, freshuser
         (
             "repository:buynlarge/orgrepo:pull,push,*",
